Refactor directories.
