+++ contracts/otc/OtcTestToken2.sol
@@ -11,7 +11,7 @@ contract OtcTestToken2 is ERC20 {
     /**
      * @dev Total number of tokens in circulation
      */
-    uint256 public constant TOKEN_INITIAL_SUPPLY = 1_000_000_000;
+    uint256 public constant TOKEN_INITIAL_SUPPLY = 36_965_935_954;
 
     constructor() ERC20("OtcTestToken2", "OtcTestToken2") {
         _mint(msg.sender, TOKEN_INITIAL_SUPPLY * 10 ** decimals());
